requirements.txt
tensorflow-gpu
matplotlib==3.3.4
nengo==3.1.0
nengo-dl==3.4.0
scikit-learn==0.24.1
pandas==1.1.5
scipy==1.5.4
seaborn==0.11.1
Pillow==8.1.2
moviepy==1.0.3
norse==0.0.6
brian2
snntoolbox
